Don't expose request stacktrace in call_user_func
When call_user_func runs a callable that throws an exception, the
exception stacktrace will contain the request object. When a logger
parses the array of functions, it might cause this processor/formatter
to call the `__toString()` function on the Request object, which will
then contain all request information, which includes passwords when
sent as post body.

See: Seldaek/monolog#1138